Factor out metadata
